-Added Real World Scale option. Using this, the models will export into a real-world scale! (Currently only works on Lightwave and OBJ exporters.) -Lightwave Exporter:  -Will now ask if you want to overwrite model files.  -Scene files will no longer be generated if the object doesn't export.  -Added Always Export Scene option. When enabled, this will export the scene file, even if the object doesn't export for some reason.  -All exported items obey the Real World Scale option. -OBJ Exporter:  -Now exports attached models! (Helms, Weapons and Shoulders)  -Now obeys the Preserve Directory option!  -Image filenames now work just like the Lightwave exporter's image filenames.  -All exported items obey the Real World Scale option. -DEV WORK:  -Massive, and I do mean MASSIVE 64-bit updates.  -Moved Mesh and Slot names to the modelexport.h  -Removed fn from GetM2TextureName, as it wasn't needed.  -Fixed MakeModelFaceForwards, and changed the 2nd option to flip the Z axis, instead of the X axis. If your model comes out backwards when using this, set flipZ to true.

+++ src/RenderTexture.cpp
@@ -99,7 +99,7 @@ void RenderTexture::Init(HWND hWnd, int width, int height, bool fboMode)
 		int curMax=0;
 
 		// Make sure its square and a power-of-two
-		for (int i=5; !end; i++) {
+		for (size_t i=5; !end; i++) {
 			int iSize = powf(2,(float)i);
 			if (iSize < texSize) {
 				curMax = iSize;
